refactor: update BASE_DIR calculation for correct path resolution; modify RondeWaarde model to use float for waarde; add None check for rows in DataRepository; create package-lock.json for frontend

+++ backend/src/models/models.py
@@ -12,7 +12,7 @@ class Instellingen(BaseModel):
 class RondeWaarde(BaseModel):
     trainings_id: int
     ronde_nummer: int
-    waarde: int
+    waarde: float
     uitkomst: str
 
 class Resultaat(BaseModel):
